Use `int` for residence capacity (percentage)

+++ src/UI/PopulationPanel.h
@@ -21,22 +21,22 @@ public:
 	 * \fixme	This class/function is use to store residence capacity
 	 *			by the MapViewState. Probably not be an appropriate place.
 	 */
-	float capacity() const { return mCapacity; }
+	int capacity() const { return mCapacity; }
 
 	virtual void update() final;
 
 protected:
 
 private:
-	NAS2D::Image		mIcons;
-	NAS2D::ImageList	mSkin;
+	NAS2D::Image mIcons;
+	NAS2D::ImageList mSkin;
 
-	Population*			mPopulation = nullptr;
+	Population* mPopulation = nullptr;
 
-	int					mResidentialCapacity = 0;
+	int mResidentialCapacity = 0;
 
-	int*				mMorale = nullptr;
-	int*				mPreviousMorale = nullptr;
+	int* mMorale = nullptr;
+	int* mPreviousMorale = nullptr;
 
-	float				mCapacity = 0.0f;
+	int mCapacity = 0;
 };
